Fixed Relay logic

+++ ph.cpp
@@ -170,7 +170,7 @@ void PH::calibratePh() {
         readPhRawToBuffer();                                   // Read raw ph value and add to buffer
 
         phVol = (float)_buf[0] * 5.0 / 1024;                   // Convert to mv and store
-        setCalActual (0, (-5.70 * phVol + 21.34));             // convert millivolts to PH reading without calibration
+        setCalActual (0, ((float) - 5.70 * phVol + 21.34));           // convert millivolts to PH reading without calibration
 
         _prevPhTime == millis();  // reset PH timer
         haveFirstPoint = true;    // Flag the first point as done
@@ -180,7 +180,7 @@ void PH::calibratePh() {
         Serial.println(F("Reading Actual 2"));
         readPhRawToBuffer();                                   // Reset buffer index
         phVol = (float)_buf[1] * 5.0 / 1024;                   // Convert to mv and store
-        setCalActual(1, (-5.70 * phVol + 21.34));                    // convert millivolts to PH reading without calibration
+        setCalActual(1, ((float) - 5.70 * phVol + 21.34));                  // convert millivolts to PH reading without calibration
 
         _prevPhTime = millis();        // reset PH timer
         _phIndex = 0;                  // Reset index so we can start getting ph readings again
